requirements.txt
theano==0.8.2
numpy>=1.10.0
psutil>=4.1.0
oauth2client==3.0.0
google-api-python-client==1.5.3
PyDrive==1.0.1
pyperclip>=1.5.27
web.py>=0.38
docopt==0.6.2
pyssim==0.3
APScheduler==3.0.3
argparse==1.2.1
watchdog==0.8.3
peakutils==1.0.3
treeherder-client==3.1.0
selenium==3.0.2
geckoprofiler_controller==0.0.4
matplotlib==1.5.3
ipython==5.3.0
pandas>=0.19.2
runipy==0.1.5

python-dateutil
keyring
requests
